Route every scenario through the codec factory
The transform and decoder-MD5 scenarios now resolve platforms the
same way as encoder PSNR and decoder performance: the genio transform
pipeline classes move into codec_genio.py as BaseCodecProject projects
(pipelines verified byte-identical), the Actions enum moves to
gst_utils as GStreamerTransformActions, and the MD5 script takes the
platform from the resource and honors an optional
build_decoder_md5_checksum_command module hook. Adding a platform is
now always just codec_<family>.py plus one PLATFORM_FAMILIES entry.
